4th commit

+++ lib/view/home/all_team_page.dart
@@ -11,6 +11,22 @@ class _AllTeamsPageState extends State<AllTeamsPage> {
   final CollectionReference team = FirebaseFirestore.instance.collection(
     'teams',
   );
+  getToken() async {
+    String? token = await FirebaseMessaging.instance.getToken();
+    await FirebaseFirestore.instance
+        .collection('Users')
+        .doc(FirebaseAuth.instance.currentUser!.uid)
+        .update({'fcm_token': token});
+    log('$token');
+  }
+
+  @override
+  void initState() {
+    if (FirebaseAuth.instance.currentUser != null) {
+      getToken();
+    }
+    super.initState();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -23,13 +39,13 @@ class _AllTeamsPageState extends State<AllTeamsPage> {
 
       body: StreamBuilder(
         stream: FirebaseAuth.instance.authStateChanges(),
-        builder: (context, usrSnapshot) {
-          if (!usrSnapshot.hasData) {
+        builder: (context, userSnapshot) {
+          if (!userSnapshot.hasData || userSnapshot.data == null) {
             return Center(
               child: CircularProgressIndicator(backgroundColor: thirdColor),
             );
           }
-          final user = usrSnapshot.data!;
+          final user = userSnapshot.data!;
           return StreamBuilder<QuerySnapshot>(
             stream:
                 FirebaseFirestore.instance
@@ -47,10 +63,10 @@ class _AllTeamsPageState extends State<AllTeamsPage> {
               final teams = snapshot.data!.docs;
               return Padding(
                 padding: const EdgeInsets.only(top: 3),
-                child: ListView.builder(
+                child: AnimatedList(
                   physics: NeverScrollableScrollPhysics(),
-                  itemCount: teams.length,
-                  itemBuilder: (context, index) {
+                  initialItemCount: teams.length,
+                  itemBuilder: (context, index, anim) {
                     return Padding(
                       padding: const EdgeInsets.only(top: 3.0),
                       child: Padding(
@@ -60,51 +76,53 @@ class _AllTeamsPageState extends State<AllTeamsPage> {
                           // subtitle: Text("${teams[index]['members_list']}"),
                           trailing: IconButton(
                             onPressed: () {
-                              // final String currentUser =
-                              //     FirebaseAuth.instance.currentUser!.uid;
+                              final String currentUser =
+                                  FirebaseAuth.instance.currentUser!.uid;
 
                               Get.defaultDialog(
                                 title: "Alert",
                                 content: Text("What you want to do??"),
                                 textConfirm: 'Edit',
                                 onConfirm: () {
-                                  // if (teams[index]['createdBy'] == currentUser) {
-                                  Get.to(
-                                    EditTeam(
-                                      id: teams[index].id,
-                                      title: teams[index]['title'],
-                                      describe: teams[index]['description'],
-                                    ),
-                                  );
-                                  // } else {
-                                  // Get.defaultDialog(
-                                  //   title: "Sorry",
-                                  //   content: Text(
-                                  //     "Only Team Leader can Edit Team",
-                                  //   ),
-                                  // );
-                                  // }
+                                  if (teams[index]['createdBy'] ==
+                                      currentUser) {
+                                    Get.to(
+                                      EditTeam(
+                                        id: teams[index].id,
+                                        title: teams[index]['title'],
+                                        describe: teams[index]['description'],
+                                      ),
+                                    );
+                                  } else {
+                                    Get.defaultDialog(
+                                      title: "Sorry",
+                                      content: Text(
+                                        "Only Team Leader can Edit Team",
+                                      ),
+                                    );
+                                  }
                                 },
                                 cancel: ElevatedButton(
                                   onPressed: () {
-                                    // if (teams[index]['createdBy'] ==
-                                    //     currentUser) {
-                                    team
-                                        .doc(teams[index].id)
-                                        .delete()
-                                        .then((value) => log("Team deleted"))
-                                        .catchError(
-                                          (error) =>
-                                              log("Failed to add user: $error"),
-                                        );
-                                    // } else {
-                                    // Get.defaultDialog(
-                                    //   title: "Sorry",
-                                    //   content: Text(
-                                    //     "Only Team Leader can Delete Team",
-                                    //   ),
-                                    // );
-                                    // }
+                                    if (teams[index]['createdBy'] ==
+                                        currentUser) {
+                                      team
+                                          .doc(teams[index].id)
+                                          .delete()
+                                          .then((value) => log("Team deleted"))
+                                          .catchError(
+                                            (error) => log(
+                                              "Failed to add user: $error",
+                                            ),
+                                          );
+                                    } else {
+                                      Get.defaultDialog(
+                                        title: "Sorry",
+                                        content: Text(
+                                          "Only Team Leader can Delete Team",
+                                        ),
+                                      );
+                                    }
                                   },
                                   child: Text("Delete"),
                                 ),
@@ -114,7 +132,12 @@ class _AllTeamsPageState extends State<AllTeamsPage> {
                             icon: Icon(Icons.more_vert),
                           ),
                           onTap: () {
-                            Get.to(() => Members(teamID: teams[index].id));
+                            Get.to(
+                              () => Members(
+                                teamID: teams[index].id,
+                                teamName: teams[index]['title'],
+                              ),
+                            );
                           },
                         ),
                       ),
